Refactored name notifications to singular
