+++ app/src/main/res/layout/notification_card_view.xml
